+++ src/components/index.ts
@@ -11,3 +11,4 @@ export { default as ListCard } from "./ListCard";
 export { default as PinDetailCard } from "./PinDetailCard";
 export { default as Window } from "./Window";
 export { default as Textarea } from "./Textarea";
+export { default as InputRadio } from "./InputRadio/InputRadio";
